refactor(PathCalculator.java): 라이브러리를 도메인 에서 제거
도메인에 존재하던 라이브러리를 도메인 격리를 위해 분리

BREKAING CHANGE: 경로 조회 관련 클래스를 인스턴스로 변경
as-is:
public interface PathCalculator {
    GraphPath<Station, DefaultWeightedEdge> findShortestPath(final Station source, final Station target);
}

@Service
public class PathService {

    ...

    public PathResponse findShortestPath(final PathRequest request) {
        ...

        final PathCalculator pathCalculator = new PathCalculatorDijkstra(lines);
        final GraphPath<Station, DefaultWeightedEdge> path = pathCalculator.findShortestPath(source, target);
        final List<Station> stations = path.getVertexList();
        final int distance = (int) path.getWeight();
        ...
        return new PathResponse(stations, distance, fare);
    }
}

to-be:
subway/domain/path/Path.java:
public interface Path {
    List<Station> getVertexList();
    int getWeight();
}

subway/infrastructure/SubwayPath.java:
public class SubwayPath implements Path {

    private final GraphPath<Station, DefaultWeightedEdge> path;

    public SubwayPath(GraphPath<Station, DefaultWeightedEdge> path) {
        this.path = path;
    }

    @Override
    public List<Station> getVertexList() {
        return path.getVertexList();
    }

    @Override
    public int getWeight() {
        return (int) path.getWeight();
    }
}

/subway/domain/path/PathCalculator.java:
public interface PathCalculator {
    Path findShortestPath(final Station source, final Station target);
}

subway/infrastructure/PathCalculatorDijkstra.java:
public class PathCalculatorDijkstra implements PathCalculator {

    ...

    @Override
    public Path findShortestPath(final Station source, final Station target) {
        return new SubwayPath(dijkstraShortestPath.getPath(source, target));
    }
}

@Service
public class PathService {

    ...

    public PathResponse findShortestPath(final PathRequest request) {
        ...

        final PathCalculator pathCalculator = new PathCalculatorDijkstra(lines);
        final Path path = pathCalculator.findShortestPath(source, target);
        final List<Station> stations = path.getVertexList();
        final int distance = path.getWeight();

        ...
        return new PathResponse(stations, distance, fare);
    }
}
