+++ src/main/java/wooteco/subway/domain/path/Path.java
@@ -1,0 +1,9 @@
+package wooteco.subway.domain.path;
+
+import java.util.List;
+import wooteco.subway.domain.Station;
+
+public interface Path {
+    List<Station> getVertexList();
+    int getWeight();
+}
