+++ src/main/java/wooteco/subway/domain/path/PathCalculator.java
@@ -1,0 +1,7 @@
+package wooteco.subway.domain.path;
+
+import wooteco.subway.domain.Station;
+
+public interface PathCalculator {
+    Path findShortestPath(final Station source, final Station target);
+}
